chore: Try PDM as package manager
This package manage does *not* use virtualenvs which means all that
stack and software isn't needed and does not interfere with tox and the
test environment.

PDM requires Python 3.7 but I do not want to maintain a CI pipeline with
multiple python versions, therefore try running all tox envs with at
least Python 3.7.